some support for resolving markdown in local host
